Finished the demo page with some card examples
