+++ index.html
@@ -730,6 +730,20 @@
       </div>
     </main>
 
-    <script src="./script.js" type="module"></script>
+    <footer class="footer">
+      <button class="buttons__item buttons--accent1" id="demoBtn">
+        View Demo
+      </button>
+
+      <p class="footer__text text--medium">
+        Made with 💚 by
+        <a href="https://juanda.dev" class="links__item links--primary"
+          >Juan Daniel Martínez</a
+        >
+      </p>
+    </footer>
+
+    <script src="./script.js"></script>
+    <script src="./themeMenu.js" type="module"></script>
   </body>
 </html>
