+++ script.js
@@ -1,5 +1,3 @@
-import { theme1, theme2, theme3, theme4, theme5 } from "./themes.js";
-
 const disabledButtons = document.getElementsByClassName("buttons--disabled");
 const disabledInputs = document.getElementsByClassName("inputs--disabled");
 const badges = document.getElementsByClassName("badge");
@@ -7,100 +5,11 @@ const alerts = document.getElementsByClassName("alert");
 const alertCloseBtns = document.getElementsByClassName("alerts__close");
 const resetAlertsBtn = document.getElementById("resetAlerts");
 const tooltips = document.getElementsByClassName("tooltip");
-const themeBar = document.querySelector(".themes");
-const themeBarOffset = themeBar.offsetTop;
-const hamburger = document.querySelector(".themes__hamburger");
-
-function toggleDarkLight(theme) {
-  let bgColor, textColor, dividerColor, alpha;
-
-  switch (theme) {
-    case "light":
-      bgColor = "#f9f9f9";
-      textColor = "#24292e";
-      dividerColor = "#37a397";
-      alpha = 1;
-
-      Array(...badges).forEach((item) => item.classList.remove("dark"));
-      Array(...alerts).forEach((item) => item.classList.remove("dark"));
-      Array(...tooltips).forEach((item) => item.classList.remove("dark"));
-      break;
-
-    case "dark":
-      bgColor = "#0d1117";
-      textColor = "#f9f9f9";
-      dividerColor = "#f9f9f9";
-      alpha = 0.5;
-
-      Array(...badges).forEach((item) => item.classList.add("dark"));
-      Array(...alerts).forEach((item) => item.classList.add("dark"));
-      Array(...tooltips).forEach((item) => item.classList.add("dark"));
-      break;
-
-    default:
-      bgColor = "#f9f9f9";
-      textColor = "#24292e";
-      dividerColor = "#37a397";
-      alpha = 1;
-
-      Array(...badges).forEach((item) => item.classList.remove("dark"));
-      Array(...alerts).forEach((item) => item.classList.remove("dark"));
-      Array(...tooltips).forEach((item) => item.classList.remove("dark"));
-      break;
-  }
-
-  document
-    .querySelector("html")
-    .style.setProperty("--background_color", bgColor);
-  document.querySelector("html").style.setProperty("--text_color", textColor);
-  document
-    .querySelector("html")
-    .style.setProperty("--divider_color", dividerColor);
-  document.querySelector("html").style.setProperty("--alpha", alpha);
-}
-
-function switchTheme(theme) {
-  for (let i = 0; i < theme.length; i++) {
-    document
-      .querySelector("html")
-      .style.setProperty(`--${theme[i][0]}`, theme[i][1]);
-  }
-}
-
-function stickThemes() {
-  if (window.pageYOffset >= themeBarOffset) {
-    themeBar.classList.add("sticky");
-  } else {
-    themeBar.classList.remove("sticky");
-  }
-}
 
 window.addEventListener("load", () => {
   document
-    .querySelector(".themes__item.theme1")
-    .addEventListener("click", () => switchTheme(theme1));
-  document
-    .querySelector(".themes__item.theme2")
-    .addEventListener("click", () => switchTheme(theme2));
-  document
-    .querySelector(".themes__item.theme3")
-    .addEventListener("click", () => switchTheme(theme3));
-  document
-    .querySelector(".themes__item.theme4")
-    .addEventListener("click", () => switchTheme(theme4));
-  document
-    .querySelector(".themes__item.theme5")
-    .addEventListener("click", () => switchTheme(theme5));
-  document.querySelector(".toggle-theme").addEventListener("change", (e) => {
-    if (e.target.checked) {
-      toggleDarkLight("dark");
-    } else {
-      toggleDarkLight("light");
-    }
-  });
-  hamburger.addEventListener("click", () => {
-    document.querySelector(".themes__collapse").classList.toggle("show");
-  });
+    .getElementById("demoBtn")
+    .addEventListener("click", () => (window.location.href = "./demo.html"));
 
   Array(...disabledButtons).forEach((item) =>
     item.addEventListener("click", (e) => e.preventDefault)
@@ -119,7 +28,3 @@ window.addEventListener("load", () => {
     );
   });
 });
-
-window.onscroll = function () {
-  stickThemes();
-};
