+++ scss/_loadings.scss
@@ -2,6 +2,20 @@
   grid-column: 1 / 2;
   grid-row: 6 / 7;
 
+  &.loadings--accent1 {
+    .loadings__variant1,
+    .loadings__variant2 {
+      &__label {
+        color: var(--accent_color_1_1);
+      }
+
+      &__dot,
+      &__bar {
+        background-color: var(--accent_color_1_1);
+      }
+    }
+  }
+
   &__variant1 {
     position: relative;
     margin: 2rem auto;
@@ -31,6 +45,8 @@
 
   &__variant2 {
     text-align: center;
+    width: 300px;
+    margin: auto;
 
     &__bar {
       position: relative;
@@ -111,10 +127,6 @@
   .loadings {
     grid-column: 1 / 3;
     grid-row: 7 / 8;
-
-    &__variant2__bar {
-      margin-left: 27%;
-    }
   }
 }
 
@@ -122,10 +134,6 @@
   .loadings {
     grid-column: 1 / 2;
     grid-row: 9 / 10;
-
-    &__variant2__bar {
-      margin-left: 17%;
-    }
   }
 }
