+++ scss/_variables.scss
@@ -2,7 +2,6 @@
   --font_family: "Poppins";
   --text_color: #24292e;
   --background_color: #f9f9f9;
-  --divider_color: #37a397;
   --alpha: 1;
 
   // State Colors
