+++ scss/demo.scss
@@ -1,0 +1,135 @@
+.main {
+  .demo {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-around;
+  }
+
+  .loadings__variant2 {
+    margin-top: 3rem;
+  }
+}
+
+.card {
+  margin-bottom: 2rem;
+  width: 300px;
+  height: 425px;
+  border-radius: 20px;
+  -webkit-border-radius: 20px;
+  -moz-border-radius: 20px;
+  -ms-border-radius: 20px;
+  -o-border-radius: 20px;
+  box-shadow: 3px 3px 20px 0 rgba(0, 0, 0, 0.3);
+
+  &.card--image {
+    position: relative;
+    background: url("https://images.unsplash.com/photo-1619376296983-4e1a253ab269?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=1000&q=80");
+    background-size: cover;
+    background-position: center;
+
+    &:hover .card__button {
+      height: 47px;
+
+      &__item {
+        padding-top: 0.7rem;
+      }
+    }
+
+    .card__button {
+      position: absolute;
+      bottom: 0;
+      left: 0;
+      width: 100%;
+      height: 0px;
+      transition: all 0.3s ease-in-out;
+      -webkit-transition: all 0.3s ease-in-out;
+      -moz-transition: all 0.3s ease-in-out;
+      -ms-transition: all 0.3s ease-in-out;
+      -o-transition: all 0.3s ease-in-out;
+
+      &__item {
+        display: block;
+        padding-top: 0;
+        width: 100%;
+        height: 100%;
+        color: #ffffff;
+        text-align: center;
+        font-weight: 700;
+        text-decoration: none;
+        background-color: var(--accent_color_1_1);
+        overflow: hidden;
+        border-radius: 0 0 20px 20px;
+        -webkit-border-radius: 0 0 20px 20px;
+        -moz-border-radius: 0 0 20px 20px;
+        -ms-border-radius: 0 0 20px 20px;
+        -o-border-radius: 0 0 20px 20px;
+      }
+    }
+  }
+
+  &__image {
+    margin-bottom: 1rem;
+    width: 100%;
+    height: 200px;
+
+    img {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+      object-position: left;
+      border-radius: 20px 20px 0 0;
+      -webkit-border-radius: 20px 20px 0 0;
+      -moz-border-radius: 20px 20px 0 0;
+      -ms-border-radius: 20px 20px 0 0;
+      -o-border-radius: 20px 20px 0 0;
+    }
+  }
+
+  &__info {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    grid-template-rows: repeat(2, auto);
+    padding: 0 1rem 1rem 1rem;
+
+    &__item {
+      &__title {
+        margin-bottom: 0.5rem;
+        text-transform: uppercase;
+      }
+
+      &.col-1 {
+        grid-column: 1 / 2;
+        grid-row: 1 / 2;
+      }
+
+      &.col-2 {
+        grid-column: 2 / 3;
+        grid-row: 1 / 2;
+      }
+
+      &.col-1,
+      &.col-2 {
+        text-align: center;
+
+        .text--light {
+          font-size: 0.8rem;
+          color: var(--neutral_color_4);
+        }
+      }
+
+      &.col-3 {
+        grid-column: 1 / 3;
+        grid-row: 2 / 3;
+        margin-top: 0.5rem;
+
+        .buttons__item {
+          display: block;
+          margin-top: 1rem;
+          margin-left: auto;
+          margin-right: auto;
+          text-decoration: none;
+        }
+      }
+    }
+  }
+}
